long * -> int * except for seeds
git-svn-id: https://svn.r-project.org/R/branches/R-0-65-patches@5805 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/appl/chisqsim.c
@@ -11,10 +11,10 @@
 #include "S.h"
 
 static void
-rcont2(long *nrow, long *ncol, long *nrowt, long *ncolt, long *ntotal,
-       double *fact, long *jwork, long *matrix) 
+rcont2(int *nrow, int *ncol, int *nrowt, int *ncolt, int *ntotal,
+       double *fact, int *jwork, int *matrix) 
 {
-    long nlmp, j, l, m, ia, ib, ic, jc, id, ie, ii, nrowtl, iap, idp,
+    int nlmp, j, l, m, ia, ib, ic, jc, id, ie, ii, nrowtl, iap, idp,
 	igp, ihp, iip, nll, nlm, nrowm, ncolm, lsm, lsp;
     double x, y, dummy, sumprb;
 
@@ -151,12 +151,12 @@ L190:
    */
 
 void
-chisqsim(long *nrow, long *ncol, long *nrowt, long *ncolt, long *n,
-	 long *b, double *expected, long *observed, double *fact,
-	 long *jwork, double *results)
+chisqsim(int *nrow, int *ncol, int *nrowt, int *ncolt, int *n,
+	 int *b, double *expected, int *observed, double *fact,
+	 int *jwork, double *results)
 {
     /* Local variables */
-    long i, j, iter;
+    int i, j, iter;
     double chi, e, o, x;
 
     /* Calculate log-factorials */
